requirements.txt
boto3==1.40.59
botocore==1.40.59
contourpy==1.3.2
cycler==0.12.1
fonttools==4.60.1
jmespath==1.0.1
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.2.6
packaging==25.0
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
s3transfer==0.14.0
six==1.17.0
urllib3==2.5.0
